+++ MicronetHardwareLibrary/app/src/main/jni/api.cpp
@@ -289,11 +289,12 @@ int set_rtc_reg_dbg(int * fd, uint8_t address, uint8_t data)
 	return set_command(fd, req, sizeof(req));
 }
 
-int check_rtc_battery(int * fd, int * battery_state)
+int check_rtc_battery(int * fd, uint8_t * battery_state)
 {
 	uint8_t address = RTC_FLAGS_ADDR;
 	uint8_t flags = 0;
     int result = get_rtc_reg_dbg(fd, address, &flags);
+    // If this statement is true, then that means the RTC Battery is bad or not present.
 	if (flags & 0x10)
 	{
         *battery_state = 0;
